+++ apps/playground/src/atom/env-config.ts
@@ -1,9 +1,9 @@
 import { atom } from "jotai"
 import { atomWithStorage } from "jotai/utils"
 
-type WaypointOrigin = "https://waypoint.roninchain.com" | "https://id.skymavis.one"
+type WaypointOrigin = "https://id.skymavis.com" | "https://id.skymavis.one"
 const CLIENT_ID_MAP: Record<WaypointOrigin, string> = {
-  ["https://waypoint.roninchain.com"]: "0e188f93-b419-4b0f-8df4-0f976da91ee6",
+  ["https://id.skymavis.com"]: "0e188f93-b419-4b0f-8df4-0f976da91ee6",
   ["https://id.skymavis.one"]: "5cf4daa9-e7ff-478b-a96c-1a9d46c916ca",
 }
 
@@ -21,8 +21,8 @@ export const waypointConfigAtom = atom(get => {
 
 export const switchWaypointConfigAtom = atom(null, async (get, set) => {
   set(_waypointOriginAtom, current => {
-    return current !== "https://waypoint.roninchain.com"
-      ? "https://waypoint.roninchain.com"
+    return current !== "https://id.skymavis.com"
+      ? "https://id.skymavis.com"
       : "https://id.skymavis.one"
   })
 })